pyproject.toml
[tool.poetry]
name = "levibot"
version = "0.1.1"
description = ""
authors = ["lev"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py-self" = "^1.9.2"
toml = "^0.10.2"
jishaku = "^2.4.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
levibot = "levibot.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
